+++ xmlunit-assertj/src/test/java/org/xmlunit/assertj/MultipleNodeAssertTest.java
@@ -33,7 +33,7 @@ public class MultipleNodeAssertTest {
     }
 
     @Test
-    public void testContainsAnyNodeHavingXPath_shouldFailed() {
+    public void testContainsAnyNodeHavingXPath_shouldFail() {
 
         thrown.expectAssertionError(format("Expecting:%nany node in set have XPath: <self::node()[@attrA]>"));
 
@@ -49,6 +49,28 @@ public class MultipleNodeAssertTest {
                 "</feed>";
 
         assertThat(xml)
+                .hasXPath("/feed/entry")
+                .containsAnyNodeHavingXPath("self::node()[@attrA]");
+    }
+
+    @Test
+    public void testContainsAnyNodeHavingXPath_shouldUseCustomFailMessage() {
+
+        thrown.expectAssertionError("Alarm alarm!");
+
+        String xml = "<?xml version=\"1.0\" encoding=\"UTF-8\"?>" +
+                "<feed>" +
+                "   <title>title</title>" +
+                "   <entry attr=\"value\">" +
+                "       <title>title1</title>" +
+                "   </entry>" +
+                "   <entry attr2=\"value2\">" +
+                "       <title>title1</title>" +
+                "   </entry>" +
+                "</feed>";
+
+        assertThat(xml)
+                .withFailMessage("Alarm alarm!")
                 .hasXPath("/feed/entry")
                 .containsAnyNodeHavingXPath("self::node()[@attrA]");
     }
@@ -74,7 +96,7 @@ public class MultipleNodeAssertTest {
     }
 
     @Test
-    public void testContainsAllNodesHavingXPath_shouldFailed() {
+    public void testContainsAllNodesHavingXPath_shouldFail() {
 
         thrown.expectAssertionError("check node at index 2");
         thrown.expectAssertionError(format("Expecting:%n <entry>%nto have XPath: <./title>"));
@@ -99,6 +121,31 @@ public class MultipleNodeAssertTest {
     }
 
     @Test
+    public void testContainsAllNodesHavingXPath_shouldUseCustomFailMessage() {
+
+        thrown.expectAssertionError("Alarm alarm!");
+
+        String xml = "<?xml version=\"1.0\" encoding=\"UTF-8\"?>" +
+                "<feed>" +
+                "   <title>title</title>" +
+                "   <entry>" +
+                "       <title>title1</title>" +
+                "   </entry>" +
+                "   <entry>" +
+                "       <title>title1</title>" +
+                "   </entry>" +
+                "   <entry>" +
+                "       <description>description</description>" +
+                "   </entry>" +
+                "</feed>";
+
+        assertThat(xml)
+                .withFailMessage("Alarm alarm!")
+                .hasXPath("/feed/entry")
+                .containsAllNodesHavingXPath("./title");
+    }
+
+    @Test
     public void testExtractingAttribute_shouldPass() {
 
         String xml = "<?xml version=\"1.0\" encoding=\"UTF-8\"?>" +
@@ -114,6 +161,47 @@ public class MultipleNodeAssertTest {
                 .nodesByXPath("/feed/entry")
                 .extractingAttribute("attr1")
                 .containsExactly("value1", "value2", null, "value4");
+    }
+
+    @Test
+    public void testExtractingAttribute_shouldFail() {
+
+        thrown.expectAssertionError("[XPath \"/feed/entry\" evaluated to node set] ");
+
+        String xml = "<?xml version=\"1.0\" encoding=\"UTF-8\"?>" +
+                "<feed>" +
+                "   <title>title</title>" +
+                "   <entry attr1=\"value1\" />" +
+                "   <entry attr1=\"value2\"/>" +
+                "   <entry />" +
+                "   <entry attr1=\"value4\" />" +
+                "</feed>";
+
+        assertThat(xml)
+                .nodesByXPath("/feed/entry")
+                .extractingAttribute("attr1")
+                .containsExactly("value1", "value2", "value3", "value4");
+    }
+
+    @Test
+    public void testExtractingAttribute_shouldUseCustomErrorMessage() {
+
+        thrown.expectAssertionError("Alarm alarm!");
+
+        String xml = "<?xml version=\"1.0\" encoding=\"UTF-8\"?>" +
+                "<feed>" +
+                "   <title>title</title>" +
+                "   <entry attr1=\"value1\" />" +
+                "   <entry attr1=\"value2\"/>" +
+                "   <entry />" +
+                "   <entry attr1=\"value4\" />" +
+                "</feed>";
+
+        assertThat(xml)
+                .withFailMessage("Alarm alarm!")
+                .nodesByXPath("/feed/entry")
+                .extractingAttribute("attr1")
+                .containsExactly("value1", "value2", "value3", "value4");
     }
 
     @Test
@@ -185,7 +273,7 @@ public class MultipleNodeAssertTest {
     }
 
     @Test
-    public void testExtractingTextMultipleNodes_shouldFailed() {
+    public void testExtractingTextMultipleNodes_shouldFail() {
 
         thrown.expectAssertionError("[XPath \"/document/h2\" evaluated to node set] ");
 
@@ -196,6 +284,24 @@ public class MultipleNodeAssertTest {
                 "</document>";
 
         assertThat(xml)
+                .nodesByXPath("/document/h2")
+                .extractingText()
+                .containsExactly("Matching header", "Header not matching");
+    }
+
+    @Test
+    public void testExtractingTextMultipleNodes_shouldUseCustomFailMessage() {
+
+        thrown.expectAssertionError("Alarm alarm!");
+
+        String xml = "<?xml version=\"1.0\" encoding=\"UTF-8\"?>" +
+                "<document>" +
+                "   <h2> Matching header</h2>" +
+                " <h2>Not matching</h2>" +
+                "</document>";
+
+        assertThat(xml)
+                .withFailMessage("Alarm alarm!")
                 .nodesByXPath("/document/h2")
                 .extractingText()
                 .containsExactly("Matching header", "Header not matching");